+++ tests/unit/components/drawer/ComponentDetailPanel.spec.js
@@ -76,6 +76,7 @@ describe('test component: Plugin Component Detail Panel', () => {
                 type: 'componentType',
               }],
             },
+            renameComponentId: jest.fn(),
           },
           draw: jest.fn(),
         },
@@ -114,6 +115,10 @@ describe('test component: Plugin Component Detail Panel', () => {
         }],
       }];
 
+      wrapper.vm.originalComponent = {
+        createAttribute: jest.fn((props) => new ComponentAttribute(props)),
+      };
+
       expect(wrapper.vm.sanitizeAttributes(attributes)).toEqual([new ComponentAttribute({
         name: 'attribute',
         value: 'value',
@@ -140,14 +145,12 @@ describe('test component: Plugin Component Detail Panel', () => {
 
       wrapper.vm.submit();
 
-      expect(wrapper.vm.originalComponent.id).toEqual(wrapper.vm.selectedComponentId);
       expect(wrapper.vm.originalComponent.attributes)
         .toEqual(wrapper.vm.selectedComponentAttributes);
       expect(wrapper.vm.isVisible).toEqual(false);
 
       wrapper.vm.submit();
 
-      expect(wrapper.vm.originalComponent.id).toEqual(wrapper.vm.selectedComponentId);
       expect(wrapper.vm.originalComponent.attributes)
         .toEqual(wrapper.vm.selectedComponentAttributes);
       expect(wrapper.vm.isVisible).toEqual(false);
@@ -204,6 +207,7 @@ describe('test component: Plugin Component Detail Panel', () => {
         definition: {
           definedAttributes: [definition],
         },
+        createAttribute: jest.fn((props) => new ComponentAttribute(props)),
       };
 
       expect(wrapper.vm.getReferencedAttributes(component)).toEqual([{
@@ -337,8 +341,12 @@ describe('test component: Plugin Component Detail Panel', () => {
   });
 
   describe('Test function: getAttributeByDefinition', () => {
+    const component = {
+      ...new Component(),
+      createAttribute: jest.fn((props) => new ComponentAttribute(props)),
+    };
+
     it('should return attribute with null value when type is not Object', () => {
-      const component = new Component();
       const definition = new ComponentAttributeDefinition({
         name: 'test',
         type: 'String',
@@ -354,7 +362,6 @@ describe('test component: Plugin Component Detail Panel', () => {
     });
 
     it('should return attribute with array value when type is Object', () => {
-      const component = new Component();
       const definition = new ComponentAttributeDefinition({
         name: 'test',
         type: 'Object',