+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Display buttons for each OIDC provider and perform user authentication if the global configuration file is declared.
 * Draw page:
   * Add variables panel in left drawer.
+  * Rename component id.
 * Splash screen.
 
 ### Changed
@@ -44,9 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Draw page:
   * Rework on left drawer.
 * Plugin management:
-  * Work with [terrator-plugin](https://github.com/ditrit/terrator-plugin/tree/0.4.1) version 0.4.1.
-  * Work with [plugin-core](https://github.com/ditrit/leto-modelizer-plugin-core/tree/0.18.0) version 0.18.0.
-  * Work with [githubator-plugin](https://github.com/ditrit/githubator-plugin/tree/0.2.2) version 0.2.2.
+  * Work with [terrator-plugin](https://github.com/ditrit/terrator-plugin/tree/0.5.0) version 0.5.0.
+  * Work with [plugin-core](https://github.com/ditrit/leto-modelizer-plugin-core/tree/0.19.0) version 0.19.0.
+  * Work with [githubator-plugin](https://github.com/ditrit/githubator-plugin/tree/0.2.3) version 0.2.3.
 * Improve performance of the automatic functionnal tests.
 
 ### Fixed